# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CrossUniversityLibrary
  description: OpenAPI documentation for RESTful Service CrossUniversityLibrary
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Localhost Server URL
  - url: http://25.59.220.248:8080
    description: Hamachi Server URL
security:
  - bearerAuth: []
paths:
...
  /auth/university-admin-registration:
    post:
      tags:
        - Authentication Controller
      summary: University administrator registration by global administrators
      description: >-
        This API endpoint is accessible to users with the role of global
        administrators, allowing them to register university administrators in
        the system using their email addresses. The response includes a
        CredentialDTO containing the registration information(email and
        generated password). After successful registration, the system sends an
        email to the university administrator containing login credentials.
      operationId: universityAdminRegistration
      parameters:
        - name: email
          in: query
          description: Email address of the university admin to be registered
          required: true
          schema:
            type: string
      responses:
        '201':
          description: >-
            The resource has been successfully created. The server has processed
            the request to create a new resource, and the operation was
            completed without errors.
          content:
            application/json:
              example:
                email: email_example@kpi.ua
                password: password_example
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /auth/teacher-registration:
    post:
      tags:
        - Authentication Controller
      summary: Teacher registration by university administrators
      description: >-
        This API endpoint is accessible to users with the role of university
        administrators, specifically allowing them to register teachers in the
        system using their email addresses. The response includes a
        CredentialDTO containing the registration information(email and
        generated password). After successful registration, the system sends an
        email to the teacher containing login credentials.
      operationId: teacherRegistration
      parameters:
        - name: email
          in: query
          description: Email address of the teacher to be registered
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                email: email_example@kpi.ua
                password: password_example
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /auth/student-registration:
    post:
      tags:
        - Authentication Controller
      summary: Student registration
      description: >-
        This API endpoint is designed for student registration, allowing new
        students to register by providing their email address and password. The
        method accepts parameters for the student's email and password. The
        response includes an AuthenticationResponse containing the
        authentication token.
      operationId: register
      parameters:
        - name: email
          in: query
          description: Email address for student registration
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: Password for student registration
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                accessToken: accessToken_example
                refreshToken: refreshToken_example
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /auth/authenticate:
    post:
      tags:
        - Authentication Controller
      summary: User authentication
      description: >-
        This API endpoint is designed for user authentication, allowing
        registered users to authenticate themselves by providing their email
        address and password. The method accepts parameters for the student's
        email and password. The response includes an AuthenticationResponse
        containing the authentication token.
      operationId: authentication
      parameters:
        - name: email
          in: query
          description: The email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                access_token: access_token_example
                refresh_token: refresh_token_example
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT authentication
      scheme: bearer
      bearerFormat: JWT